requirements.txt
pygame-ce>=2.3
websockets>=13
